requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.0.4
h5py==3.1.0
Keras==2.4.3
numpy==1.19.4
opencv-contrib-python==4.5.1.48
opencv-contrib-python-headless==4.5.1.48
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
PyYAML==5.3.1
scipy==1.5.4
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
tqdm==4.53.0
urllib3==1.26.2
virtualenv==20.1.0
whitenoise==5.2.0
